+++ gastropaysdk/src/main/java/com/inventiv/gastropaysdk/data/response/TermsAndConditionResponse.kt
@@ -1,0 +1,3 @@
+package com.inventiv.gastropaysdk.data.response
+
+data class TermsAndConditionResponse(val url: String)