+++ gastropaysdk/src/main/java/com/inventiv/gastropaysdk/shared/GastroPaySdkListener.kt
@@ -3,6 +3,6 @@ package com.inventiv.gastropaysdk.shared
 interface GastroPaySdkListener {
     fun onInitialized(isInitialized: Boolean) {}
     fun onAuthTokenReceived(authToken: String) {} //TODO will be update after refresh-token flow
-    fun onPaymentSuccess(){}
+    fun onPaymentSuccess() {}
     fun onSDKClosed() {}
 }