+++ gastropaysdk/src/main/java/com/inventiv/gastropaysdk/ui/wallet/WalletFragment.kt
@@ -15,6 +15,7 @@ import com.inventiv.gastropaysdk.repository.WalletRepositoryImp
 import com.inventiv.gastropaysdk.shared.GastroPaySdk
 import com.inventiv.gastropaysdk.ui.MainViewModel
 import com.inventiv.gastropaysdk.ui.MainViewModelFactory
+import com.inventiv.gastropaysdk.ui.settings.SettingsFragment
 import com.inventiv.gastropaysdk.utils.delegate.viewBinding
 import com.inventiv.gastropaysdk.utils.handleError
 import com.inventiv.gastropaysdk.view.GastroPaySdkToolbar
@@ -29,6 +30,9 @@ internal class WalletFragment : BaseFragment(R.layout.fragment_wallet_gastropay_
         showToolbar(true, toolbar, logo)
         toolbar.onRightIconClick {
             sharedViewModel.closeSdk()
+        }
+        toolbar.onLeftIconClick {
+            sharedViewModel.pushFragment(SettingsFragment())
         }
     }