+++ gastropaysdk/src/main/java/com/inventiv/gastropaysdk/ui/webview/WebViewViewModelFactory.kt
@@ -1,0 +1,16 @@
+package com.inventiv.gastropaysdk.ui.webview
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+@Suppress("UNCHECKED_CAST")
+internal class WebViewViewModelFactory() :
+    ViewModelProvider.NewInstanceFactory() {
+
+    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(WebViewViewModel::class.java)) {
+            return WebViewViewModel() as T
+        }
+        throw IllegalArgumentException("Unknown class name")
+    }
+}